backend/requirements.txt
pydantic<2.0
llama-index
openai
pypdf
python-dotenv
flask
flask-cors